Renamed parsers to yamlParser04 and yamlParser07
