+++ test/autoCompletion.test.ts
@@ -17,7 +17,7 @@ import * as URL from 'url';
 import fs = require('fs');
 import {JSONSchemaService} from '../src/languageservice/services/jsonSchemaService'
 import {schemaRequestService, workspaceContext}  from './testHelper';
-import { parse as parseYAML } from '../src/languageservice/parser/yamlParser';
+import { parse as parseYAML } from '../src/languageservice/parser/yamlParser04';
 import { getLineOffsets } from '../src/languageservice/utils/arrUtils';
 var assert = require('assert');
