+++ test/autoCompletion3.test.ts
@@ -12,7 +12,7 @@ import {
 import {getLanguageService} from '../src/languageservice/yamlLanguageService'
 import {JSONSchemaService} from '../src/languageservice/services/jsonSchemaService'
 import {schemaRequestService, workspaceContext}  from './testHelper';
-import { parse as parseYAML } from '../src/languageservice/parser/yamlParser';
+import { parse as parseYAML } from '../src/languageservice/parser/yamlParser04';
 import { getLineOffsets } from '../src/languageservice/utils/arrUtils';
 var assert = require('assert');
