+++ test/customTags.test.ts
@@ -17,7 +17,7 @@ import * as URL from 'url';
 import fs = require('fs');
 import {JSONSchemaService} from '../src/languageservice/services/jsonSchemaService'
 import {schemaRequestService, workspaceContext, createJSONLanguageService}  from './testHelper';
-import { parse as parseYAML } from '../src/languageservice/parser/yamlParser2';
+import { parse as parseYAML } from '../src/languageservice/parser/yamlParser07';
 var assert = require('assert');
 
 function createLanguageServiceWithCustomTags(customTags) {
